% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmatch-package.R
\name{boot_match_univariate}
\alias{boot_match_univariate}
\title{Match two groups on a single variable using bootstrapping}
\usage{
boot_match_univariate(data, y, x, caliper = 0, boot = 100, id = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{y}{a column in \code{data} with two values or a logical expression that can
be evaluated inside of \code{data} to determine groups.}

\item{x}{a column in \code{data}}

\item{caliper}{a caliper constraint to set on the matching. Only matches with
\code{abs(x[1] - x[2]) &lt;= caliper} are legal. A caliper of 0 would enforce exact
matches.}

\item{boot}{number of bootstrap samples to take. Defaults to 100.}

\item{id}{an optional column in \code{data} with a unique ID for each row in
\code{data}. Matches will be provided using these IDs instead of row numbers.}
}
\value{
the best bootstrapped match. The best match has the largest number of
matched pairs and the smallest \emph{z}-score difference on \code{x} between the two
matched subgroups.
}
\description{
Match two groups on a single variable using bootstrapping
}
