% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmatch-package.R
\docType{package}
\name{bootmatch-package}
\alias{bootmatch}
\alias{bootmatch-package}
\title{bootmatch: Bootstrap Group Matching}
\description{
Generate matchings between two groups on a covariate using bootstrapping.
}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@gmail.com}

}
\keyword{internal}
