% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmatch-package.R
\docType{package}
\name{bootmatch-package}
\alias{bootmatch}
\alias{bootmatch-package}
\title{bootmatch: Univariate Bootstrap Group Matching}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@gmail.com}

}
\keyword{internal}
